Called the updateJobsInThisCPU() in the dispatch() so the process ID of the current job could be added to the vector that holds which jobs are assigned to a specific CPU

+++ Ravos3/CPUMetrics.cpp
@@ -38,6 +38,6 @@ std::vector<int> CPUMetrics::getWhichJobsInThisCPU()
 //returns percentage of jobs in this CPU compared to all jobs executed
 int CPUMetrics::getPercentOfJobsAssigned()
 {
-	percentOfJobs = ((JobCount / TOTAL_JOBS_EXECUTED) * 100);
+	percentOfJobs = ((static_cast<double>(JobCount) / TOTAL_JOBS_EXECUTED) * 100);
 	return percentOfJobs;
 }
